+++ MyOwnRayTracer/renderParameters.h
@@ -6,7 +6,7 @@ public:
 
 	bool quietMode = false;
 
-	float ratio = 16.0 / 9.0;
+	double ratio = 16.0 / 9.0;
 	unsigned int width = 256;
 	unsigned int height = static_cast<unsigned int>(width / ratio);
